DashboardService/requirements.txt
fastapi
uvicorn
pydantic[email]
python-dotenv
bcrypt
sqlalchemy
asyncmy
databases
cryptography
itsdangerous
python-jose
PyJWT